move slots availability check in validators
