+++ app/transactions/admin/enrollments/update_transaction.rb
@@ -32,7 +32,6 @@ module Admin
       def validation(input) # rubocop:disable Metrics/AbcSize
         enrollment = input[:enrollment]
         enrollment.check_if_enrollment_is_matching_the_mission_s_timeslots
-        enrollment.check_slots_availability_for_regulated_mission
         return Failure(enrollment.errors.values.flatten[0]) if enrollment.errors.present?
 
         Success(input)
